rearrange files and add page actions
